# analytics-service/src/main/java/ru/retailhub/analytics/controller/AnalyticsController.java
package ru.retailhub.analytics.controller;

import lombok.RequiredArgsConstructor;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import ru.retailhub.analytics.controller.dto.ConsultantStatsResponse;
import ru.retailhub.analytics.controller.dto.DashboardResponse;
import ru.retailhub.analytics.controller.dto.RequestHistoryResponse;
import ru.retailhub.analytics.service.AnalyticsService;

import java.time.OffsetDateTime;
import java.util.List;
import java.util.UUID;

@RestController
@RequestMapping("/api/v1/analytics")
@RequiredArgsConstructor
public class AnalyticsController {

    private final AnalyticsService analyticsService;

    @GetMapping("/dashboard")
    public ResponseEntity<DashboardResponse> dashboard(
            @RequestHeader("X-Store-Id") UUID storeId,
            @RequestParam(defaultValue = "today") String period) {
        return ResponseEntity.ok(analyticsService.getDashboard(storeId, period));
    }

    @GetMapping("/consultants")
    public ResponseEntity<List<ConsultantStatsResponse>> consultantStats(
            @RequestHeader("X-Store-Id") UUID storeId,
            @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) OffsetDateTime dateFrom,
            @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) OffsetDateTime dateTo) {
        return ResponseEntity.ok(analyticsService.getConsultantStats(storeId, dateFrom, dateTo));
    }

    @GetMapping("/consultants/{userId}")
    public ResponseEntity<ConsultantStatsResponse> consultantDetail(
            @PathVariable UUID userId,
            @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) OffsetDateTime dateFrom,
            @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) OffsetDateTime dateTo) {
        return ResponseEntity.ok(analyticsService.getConsultantDetail(userId, dateFrom, dateTo));
    }

    @GetMapping("/requests")
    public ResponseEntity<RequestHistoryResponse> requestHistory(
            @RequestHeader("X-Store-Id") UUID storeId,
            @RequestParam(required = false) UUID departmentId,
            @RequestParam(required = false) String status,
            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) OffsetDateTime dateFrom,
            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) OffsetDateTime dateTo,
            @RequestParam(defaultValue = "0") int page,
            @RequestParam(defaultValue = "20") int size) {
        return ResponseEntity.ok(
                analyticsService.getRequestHistory(storeId, departmentId, status, dateFrom, dateTo, page, size));
    }
}