# api-gateway/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: RetailHub API
  description: |
    Event-Driven платформа диспетчеризации консультантов в ритейле.
    Клиент сканирует QR-код → создаётся заявка → консультант берёт заявку.

    ## Аутентификация
    Закрытые эндпоинты требуют JWT Bearer-токен:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Роли
    - **MANAGER** — управление магазином, отделами, консультантами, аналитика
    - **CONSULTANT** — приём и обработка заявок, управление сменами

    ## Публичные эндпоинты (без токена)
    - Сканирование QR-кода
    - Создание заявки
    - Polling статуса заявки клиентом
    - Напоминание консультанту
    - Смена консультанта
  version: 1.0.0
  contact:
    name: RetailHub Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: http://83.147.255.205:8180/api/v1
    description: Production (Direct IP)
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RequestStatus:
      type: string
      enum: [CREATED, WAITING, ESCALATED, ASSIGNED, COMPLETED, CANCELED]
      description: |
        CREATED  → WAITING   (> 3 мин без реакции, scheduler)
        WAITING  → ESCALATED (> 5 мин, SLA нарушен, scheduler)
        CREATED | WAITING | ESCALATED → ASSIGNED (консультант взял заявку)
        ASSIGNED → COMPLETED (консультант завершил)
        ASSIGNED → CREATED   (клиент нажал "Сменить консультанта")
        Любой, кроме COMPLETED/CANCELED → CANCELED (клиент отменил)

    # --- Auth ---
    LoginRequest:
      type: object
      required: [phone_number, password]
      properties:
        phone_number:
          type: string
          description: Номер телефона
          example: "+79991234567"
        password:
          type: string
          format: password
          minLength: 6
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access-токен (15 мин)
        refresh_token:
          type: string
          description: JWT refresh-токен (7 дней)
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Время жизни access-токена (секунды)
          example: 900
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    # --- User ---
...
    ServiceRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        store_id:
          type: string
          format: uuid
        department_id:
          type: string
          format: uuid
        department_name:
          type: string
        is_escalated:
          type: boolean
          description: "True, если заявка просрочена (> 5 мин) и эскалирована менеджеру"
        assigned_user:
          $ref: "#/components/schemas/AssignedConsultant"
        status:
          $ref: "#/components/schemas/RequestStatus"
        client_session_token:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        assigned_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    AssignedConsultant:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
    CreateRequestRequest:
      type: object
      required: [qr_token]
      properties:
        qr_token:
          type: string
          format: uuid
    ClientRequestView:
      type: object
      description: Вид заявки для клиента (polling)
      properties:
        id:
          type: string
          format: uuid
        client_session_token:
          type: string
          format: uuid
          description: "Секретный токен для polling. Возвращается ТОЛЬКО при создании."
        status:
          $ref: "#/components/schemas/RequestStatus"
        department_name:
          type: string
        consultant_name:
          type: string
          description: Имя консультанта (при ASSIGNED)
        can_remind:
          type: boolean
          description: true если прошло >= 1 мин - кнопка "Напомнить"
        can_reassign:
          type: boolean
          description: true если прошло >= 3 мин - кнопка "Сменить"
        created_at:
          type: string
          format: date-time
        assigned_at:
          type: string
          format: date-time

    # --- Shift ---
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        type:
          type: string
          description: "REQUEST_NEW, REQUEST_ASSIGNED, REQUEST_REMINDER, SHIFT_REMINDER"
        payload:
          type: object
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time

    # --- Device ---
    RegisterDeviceRequest:
      type: object
      required: [fcm_token]
      properties:
        fcm_token:
          type: string
          maxLength: 500
        device_info:
          type: string
          maxLength: 255
          example: "Samsung Galaxy S24, Android 15"
    Device:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fcm_token:
          type: string
        device_info:
          type: string
        created_at:
          type: string
          format: date-time

    # --- Analytics ---
...
    ConsultantStats:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        total_requests:
          type: integer
        completed_requests:
          type: integer
        avg_reaction_seconds:
          type: number
          format: double
        avg_service_seconds:
          type: number
          format: double
        total_work_minutes:
          type: integer
          description: Время на смене (минуты)
        total_busy_minutes:
          type: integer
          description: Время в BUSY (обслуживание)
        total_idle_minutes:
          type: integer
          description: Время простоя (ACTIVE, без заявок)
    ConsultantDetailedStats:
      type: object
      properties:
        consultant:
          $ref: "#/components/schemas/ConsultantStats"
        daily_breakdown:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              requests_completed:
                type: integer
              avg_reaction_seconds:
                type: number
                format: double
              work_minutes:
                type: integer

    # --- Common ---
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: REQUEST_ALREADY_ASSIGNED
        message:
          type: string
          example: Заявка уже взята другим консультантом
        timestamp:
          type: string
          format: date-time
    PageResponse:
      type: object
      properties:
        content:
          type: array
          items: {}
        page:
          type: integer
        size:
          type: integer
        total_elements:
          type: integer
        total_pages:
          type: integer

# PATHS
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Логин по номеру телефона
      description: Возвращает access (15 мин) + refresh (7 дней) токены.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Токены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Обновить access-токен
      description: Мобилка автоматически обновляет токен без повторного логина.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Новые токены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Refresh-токен невалиден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /requests:
    post:
      tags: [Requests]
      summary: Создать заявку (публичный, клиент через QR)
      description: |
        Флоу:
        1. GET /qr-codes/scan/{token} — инфо об отделе
        2. POST /requests — создание заявки
        3. GET /requests/{id}?session={token} — polling статуса
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestRequest"
      responses:
        "201":
          description: Заявка создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientRequestView"
        "400":
          description: QR невалиден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [Requests]
      summary: Список заявок (MANAGER/CONSULTANT)
      description: |
        Консультант видит CREATED-заявки своих отделов.
        Менеджер видит все заявки магазина.
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RequestStatus"
        - name: department_id
          in: query
          schema:
            type: string
            format: uuid
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Заявки
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageResponse"
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/ServiceRequest"
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Inbox уведомлений
      parameters:
        - name: is_read
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Уведомления
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageResponse"
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    post:
      tags: [Notifications]
      summary: Отметить прочитанным
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Отмечено

  # --- DEVICES ---
  /devices:
    post:
      tags: [Devices]
      summary: Регистрация FCM-токена
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceRequest"
      responses:
        "201":
          description: Зарегистрировано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{deviceId}:
    delete:
      tags: [Devices]
      summary: Удалить устройство (при логауте)
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Удалено

  # --- ANALYTICS (MANAGER) ---
...
  /analytics/consultants:
    get:
      tags: [Analytics]
      summary: Статистика по консультантам
      description: Заявки, время работы/простоя, реакции за период.
      parameters:
        - name: date_from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Статистика
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsultantStats"
  /analytics/consultants/{userId}:
    get:
      tags: [Analytics]
      summary: Детальная статистика консультанта
      description: С разбивкой по дням.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: date_from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Детальная статистика
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultantDetailedStats"
  /analytics/requests:
    get:
      tags: [Analytics]
      summary: История заявок
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RequestStatus"
        - name: department_id
          in: query
          schema:
            type: string
            format: uuid
        - name: assigned_user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Заявки
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageResponse"
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/ServiceRequest"